requirements.txt
backports.csv==1.0.7
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
django-crispy-forms==1.8.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-reversion==3.0.4
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.18.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupSafe==1.1.1
mistune==0.8.4
odfpy==1.4.0
openpyxl==3.0.0
Pillow==5.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.6
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
